requirements.txt
gdm-concordia
mlflow>=2.8.0
wandb>=0.15.0
ray[tune]>=2.6.0
hydra-core>=1.3.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
omegaconf>=2.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0